+++ src/com/juniperbrew/minimus/components/Rotation.java
@@ -4,11 +4,11 @@ package com.juniperbrew.minimus.components;
  * Created by Juniperbrew on 26.6.2015.
  */
 public class Rotation extends Component {
-    public int degrees;
+    public float degrees;
 
     public Rotation(){}
 
-    public Rotation(int degrees){
+    public Rotation(float degrees){
         this.degrees = degrees;
     }
 }
